Add the test file, add interaction.js for easy testing
