+++ package.json
@@ -4,6 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
+    "test": "jest",
     "build": "webpack",
     "start": "webpack serve --open"
   },
@@ -19,12 +20,14 @@
   },
   "homepage": "https://github.com/Kaweesi-Matia/To-do-list_APP#readme",
   "devDependencies": {
+    "@babel/preset-env": "^7.19.4",
     "babel-eslint": "^10.1.0",
     "css-loader": "^6.7.1",
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.26.0",
     "hint": "^7.1.2",
+    "jest": "^29.1.2",
     "style-loader": "^3.3.1",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
